Major overhaul for the multi-column render engine.
Entirely overhauled the parsing and render engine in order to fix multiple issues with previous way of handling the multi-column regions. This method is more inline with my original plan, it just took more time for it to fall.

There are still some bugs but it is in a stable enough place for now.